+++ src/App.js
@@ -418,12 +418,10 @@ class App extends Component {
         this.setState({
           plan: response.data.plan
         });
-        const addresses = window.Pelcro.address.list();
+
         const isAuthenticated = window.Pelcro.user.isAuthenticated();
         if (!isAuthenticated) {
           this.setView("register");
-        } else if (addresses && addresses.length) {
-          this.setView("payment");
         } else {
           this.setView("address");
         }
@@ -432,7 +430,7 @@ class App extends Component {
   };
 
   handleAfterRegistrationLogic = () => {
-    const { product, order, giftCode, isGift, site } = this.state;
+    const { product, order, giftCode, isGift } = this.state;
 
     ReactGA.event({
       category: "ACTIONS",
@@ -462,7 +460,7 @@ class App extends Component {
     }
 
     if (product) {
-      if (product.address_required || site.taxes_enabled) {
+      if (product.address_required) {
         return this.setView("address");
       } else {
         return this.setView("payment");
@@ -615,10 +613,7 @@ class App extends Component {
                   lastName: giftRecipient.lastName
                 });
 
-                if (
-                  this.state.product.address_required ||
-                  this.state.site.taxes_enabled
-                ) {
+                if (this.state.product.address_required) {
                   this.setView("address");
                 } else {
                   this.setView("payment");
@@ -792,17 +787,11 @@ class App extends Component {
               onSuccess={(items) => {
                 this.setOrder(items);
 
-                if (window.Pelcro.user.isAuthenticated()) {
-                  if (
-                    !window?.Pelcro?.user?.read()?.addresses?.length
-                  ) {
-                    return this.setView("address");
-                  } else {
-                    this.setView("orderCreate");
-                  }
-                } else {
-                  this.setView("register");
+                if (!window.Pelcro.user.isAuthenticated()) {
+                  return this.setView("register");
                 }
+
+                return this.setView("address");
               }}
             />
           )}